Add a button to load a map from the DB

+++ src/main/java/net/defade/amestify/utils/ProgressTracker.java
@@ -15,7 +15,7 @@ public class ProgressTracker {
         this.message = message;
     }
 
-    public void increment(int amount, String message) {
+    public void increment(long amount, String message) {
         current += amount;
         this.message = message;
     }
